Rename ot2talker_api ot2_client
