+++ src/ot2_client/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ot2talker_api</name>
+  <name>ot2_client</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="alan.linghao.wang@gmail.com">nerra</maintainer>
