+++ src/ot2_client/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/ot2_client
+[install]
+install-scripts=$base/lib/ot2_client
