+++ src/ot2talker/ot2talker/protocol_manager.py
@@ -18,10 +18,10 @@ from ot2_workcell_manager_client.worker_info_api import _get_node_info, _get_nod
 from random import random
 from arm_client.transfer_api import *
 from arm_client.transfer_api import _load_transfer
-#TODO: import ot2talker_api
-from ot2talker_api.publish_ot2_state_api import *
-from ot2talker_api.publish_ot2_state_api import _update_ot2_state
-from ot2talker_api.load_run_api import *
+#TODO: import ot2_client
+from ot2_client.publish_ot2_state_api import *
+from ot2_client.publish_ot2_state_api import _update_ot2_state
+from ot2_client.load_run_api import *
 
 class OT2ProtocolManager(Node):
