+++ src/ot2talker/package.xml
@@ -15,7 +15,7 @@
   <exec_depend>rostalker2interface</exec_depend>
   <exec_depend>ot2_workcell_manager_client</exec_depend>
   <exec_depend>arm_client</exec_depend>
-  <exec_depend>ot2talker_api</exec_depend>
+  <exec_depend>ot2_client</exec_depend>
 
   <export>
     <build_type>ament_python</build_type>
