--- src/ot2talker_api/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/ot2talker_api
-[install]
-install-scripts=$base/lib/ot2talker_api
